rf-engine/requirements.txt
fastapi
uvicorn
numpy
requests
msgpack
mercantile
Pillow
redis
scipy
<<<<<<< Updated upstream
=======
itmlogic
celery
sse-starlette
pymoo
networkx
scikit-learn
numba
>>>>>>> Stashed changes
